+++ src/components/index.js
@@ -1,5 +1,6 @@
 import Vue from 'vue';
 
+import StButtonGroup from './button-group/index.vue';
 import StButton from './button/index.vue';
 import StCheckbox from './checkbox/index.vue';
 import StCol from './col/index.vue';
@@ -9,13 +10,14 @@ import StRow from './row/index.vue';
 import StTextarea from './textarea/index.vue';
 
 const Components = {
+  StButtonGroup,
   StButton,
+  StCheckbox,
+  StCol,
   StIcon,
   StInput,
-  StTextarea,
   StRow,
-  StCol,
-  StCheckbox,
+  StTextarea,
 };
 
 Object.keys(Components).forEach((name) => {